ref: fix merge conflicts
