+++ changelog.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## [1.18.2 - 1.1.0]
+## [1.19 - 1.1.0]
 
-* port: 1.18.2
+* port: 1.19
 
 ## [Fix]
