+++ gradle.properties
@@ -8,10 +8,10 @@ mod_group=com.troblecodings
 mod_version=1.1.0
 
 # Minecraft
-minecraft_version=1.18.2
+minecraft_version=1.19
 
 # Forge
-forge_version=40.3.12
+forge_version=41.1.0
 loader=forge
 
 # Java
