+++ src/main/java/com/troblecodings/invisiblelights/init/ILInit.java
@@ -13,13 +13,13 @@ import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Block;
-import net.minecraftforge.event.world.BlockEvent.BreakEvent;
+import net.minecraftforge.event.level.BlockEvent.BreakEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 
-public final class ILInit {
+public class ILInit {
 
     public static final DeferredRegister<Block> BLOCKS =
             DeferredRegister.create(ForgeRegistries.BLOCKS, InvisibleLightsMain.MODID);
@@ -64,9 +64,6 @@ public final class ILInit {
     public static final RegistryObject<Block> BLOCKER = register("blocker", BlockLightBlocker::new);
     public static final RegistryObject<BlockGhostGlowstone> GHOST_GLOWSTONE =
             register("ghostglowstone", BlockGhostGlowstone::new);
-
-    private ILInit() {
-    }
 
     private static RegistryObject<Block> registerLight(final String name, final int level) {
         return register(name, () -> new BlockInvisibleLight(level));
